requirements.txt
ultralytics
pandas
tqdm
Pillow
opencv-python
matplotlib